API module vs modules resolved
I'm making the API more useful, and also reducing space in general. At the same time, I'm working on the new API Docs, that'll be uploaded temporarily in my fork, until we can fully implement it. Of course, Docs will be available for everyone in here as well. More updates soon!